+++ .ci-pipelines/ci-common-defs.sh
@@ -16,6 +16,7 @@ F90_feuler
 F90_lsode
 F90_radau
 F90_rk
+F90_rkadj
 F90_rktlm
 F90_ros
 F90_rosadj
@@ -27,7 +28,9 @@ F90_rostlm
 F90_ros_upcase
 F90_saprc_2006
 F90_sd
+F90_sd4
 F90_sdadj
+F90_sdtlm
 F90_seulex
 F90_small_strato
 "