+++ int/dvode.f90
@@ -4,7 +4,7 @@ MODULE KPP_ROOT_Integrator
   USE KPP_ROOT_Global
   USE KPP_ROOT_Parameters
   USE KPP_ROOT_JacobianSP
-  USE KPP_ROOT_LinearAlgebra, ONLY: KppDecomp, KppSolve, Set2zero, WLAMCH
+  USE KPP_ROOT_LinearAlgebra, ONLY: KppDecomp, KppSolve
   
   IMPLICIT NONE
   PUBLIC